# Logs: Performance and safety enhancements