… src/OpenTelemetry/.publicApi/net461/PublicAPI.Unshipped.txt
… src/OpenTelemetry/.publicApi/netstandard2.0/PublicAPI.Unshipped.txt
… src/OpenTelemetry/Logs/IInlineLogProcessor.cs
… src/OpenTelemetry/Logs/InlineLogProcessor{T}.cs
… src/OpenTelemetry/Logs/LogConvertingProcessor.cs
… src/OpenTelemetry/Logs/LogRecord.cs
… src/OpenTelemetry/Logs/LogRecordProcessor.cs
… src/OpenTelemetry/Logs/LogRecordScopeCallback.cs
… src/OpenTelemetry/Logs/LogRecordState.cs
… src/OpenTelemetry/Logs/LogRecordStruct.cs
… src/OpenTelemetry/Logs/OpenTelemetryLoggerOptions.cs
… src/OpenTelemetry/Logs/OpenTelemetryLoggerProvider.cs
… src/OpenTelemetry/SimpleExportProcessor.cs
… test/Benchmarks/Exporter/OtlpGrpcLogExporterBenchmarks.cs
… test/Benchmarks/Logs/LogScopeBenchmarks.cs
… test/OpenTelemetry.Tests/Logs/LogRecordTest.cs