Fixes #2905
Relates to #2932

## Goals

* SDK converts logs to into `OpenTelemetry.Logs.LogRecord` instances. During export `OtlpLogExporter` converts those instances into `Opentelemetry.Proto.Logs.V1.LogRecord` instances. In essence we are double buffering log data. It would be more efficient to convert the logs directly into `Opentelemetry.Proto.Logs.V1.LogRecord` instances for export.
* When SDK is converting logs into `OpenTelemetry.Logs.LogRecord` instances it doesn't know what the exporter will do with them. It has to be defensive. In the case of batching, it has to buffer data so it can be accessed safely later on. It has to buffer everything. It would be more efficient for SDK to only pay (cpu time, memory allocations) for what is needed.
* There is currently no way to make an allocation-free log exporter.

## Changes

SDK `OpenTelemetryLoggerProvider` now supports registration of a "converting processor chain" which is a conversion function + processor which accepts the results of the conversion (instead of `LogRecord`). This allows users to cut out the intermediary `LogRecord` type and only pay for what they need. Existing `LogRecord` processors will continue to work. Internally this is done with a non-generic `ILogProcessor` interface. In a follow-up we can also utilize this interface to support registration of zero-allocation log exporters.

## `OtlpLogExporter` Benchmarks

### Before

|                    Method | NumberOfLogs |      Mean |     Error |    StdDev |   Gen 0 | Allocated |
|-------------------------- |------------- |----------:|----------:|----------:|--------:|----------:|
| OtlpLogExporterr_Batching |           10 |  7.984 us | 0.1549 us | 0.1784 us |  3.7537 |     15 KB |
| OtlpLogExporterr_Batching |           50 | 40.955 us | 0.7468 us | 1.1178 us | 18.7378 |     77 KB |
| OtlpLogExporterr_Batching |          100 | 81.649 us | 1.5015 us | 1.4045 us | 37.4756 |    153 KB |

### After

|                    Method | NumberOfLogs |      Mean |     Error |    StdDev |   Gen 0 | Allocated |
|-------------------------- |------------- |----------:|----------:|----------:|--------:|----------:|
| OtlpLogExporterr_Batching |           10 |  **6.277 us** | 0.1231 us | 0.1264 us |  3.0060 |     **12 KB** |
| OtlpLogExporterr_Batching |           50 | **31.611 us** | 0.6162 us | 0.7096 us | 15.0146 |     **61 KB** |
| OtlpLogExporterr_Batching |          100 | **63.714 us** | 1.2296 us | 1.6415 us | 30.0293 |    **123 KB** |

## Example

```csharp
// Step 1: Define class to store log data.
public class MyExporterLogData
{
    public string Message { get; init; }

    // Step 2: Define conversion function to build storage class from log.
    internal static MyExporterLogData ToLogData(in LogRecordStruct log)
    {
        return new MyExporterLogData
        {
            Message = log.FormattedMessage,
        };
    }
}

// Step 3: Define exporter which accepts batch of storage class.
public class MyExporter : BaseExporter<MyExporterLogData>
{
    public override ExportResult Export(in Batch<MyExporterLogData> batch)
    {
        foreach (MyExporterLogData log in batch)
        {
            Console.WriteLine($"{log}");
        }

        return ExportResult.Success;
    }
}

// Step 4: Register converting processor chain.
public static class LoggerExtensions
{
    public static OpenTelemetryLoggerOptions AddMyExporter(this OpenTelemetryLoggerOptions options)
    {
        if (options == null)
        {
            throw new ArgumentNullException(nameof(options));
        }

        return options.AddProcessor(
            MyExporterLogData.ToLogData, // <- Conversion function.
            new BatchExportProcessor<MyExporterLogData>(new MyExporter())); // <- Processor.
    }
}
```

## Safety

* New struct to store log data `LogRecordStruct` which is `readonly ref struct`. Special type of struct (like `Span<T>`) which may not escape the stack. This is for safety. In the conversion functions users won't be allowed to copy log to heap. Or jump to another thread. This ensures the log data may only be accessed inline where it is valid. `LogRecordStruct` does not expose state `IReadOnlyList<KeyValuePair<string, object>>` only an enumerator to loop the values. This fixes the issue of deferred states being accessed when they have become invalid. Basically in the conversion function they must be looped inline so the values are captured and become safe to export later.

## Public API Changes

```csharp
// New. Conversion delegate. Must be public so it can be registered with log provider.
public delegate T LogConverter<T>(in LogRecordStruct log);

// New. Previously we used Action<LogRecordScope, TState> for looping scopes.
// Now LogRecordScope is a readonly ref struct which may not be used as
// a type parameter. To get around that, needed to define a delegate.
public delegate void LogRecordScopeCallback<TState>(LogRecordScope scope, TState state);

// New. Structure to expose state values enumerator.
public readonly ref struct LogRecordState
{
   public object State { get; }
   public Enumerator GetEnumerator();
}

// New. Structure to expose log data.
public readonly ref struct LogRecordStruct
{
   public DateTime Timestamp { get; }
   public string CategoryName { get; }
   public LogLevel LogLevel { get; }
   public EventId EventId { get; }
   public string FormattedMessage { get; }
   public Exception Exception { get; }
   public LogRecordState State { get; }
   public void ForEachScope<TState>(LogRecordScopeCallback<TState> callback, TState state);
}

// Existing
public class OpenTelemetryLoggerOptions
{
        // New. Entry point for registration of conversion function + processor.
        public OpenTelemetryLoggerOptions AddProcessor<T>(LogConverter<T> logConverter, BaseProcessor<T> processor)
            where T : class
}

// Dropped abstract from BatchExportProcessor<T>. So it can be used for whatever type the user wants to export. Made ctor public.
// Before
public abstract class BatchExportProcessor<T>
{
   protected BatchExportProcessor(...)
}
// After
public class BatchExportProcessor<T>
{
   public BatchExportProcessor(...)
}

// Dropped abstract from SimpleExportProcessor<T>. So it can be used for whatever type the user wants to export. Made ctor public.
// Before
public abstract class SimpleExportProcessor<T>
{
   protected SimpleExportProcessor(...)
}
// After
public class SimpleExportProcessor<T>
{
   public SimpleExportProcessor(...)
}
```

## Strategic Breaking Changes

I made some breaking changes. 😱 I know! I did this because it increases the safety of the solution. Impact should be very very low, if any. Only `LogExporter` authors would be impacted, and only under certain conditions. So I think it is worth doing. But not strictly necessary to achieve the goals. Looking for feedback on this.

* `readonly struct LogRecordScope` -> `readonly ref struct LogRecordScope`. Not breaking unless `LogExporter` author has copied the `LogRecordScope` to the heap, or jumped threads. That is what I'm seeking to prevent, because it is potentially unsafe. What we want to enforce/encourage is for values to be captured during the conversion, not the scope itself to be captured. Would also break if `LogExporter` author has used `LogRecordScope` as a type parameter (eg `List<LogRecordScope>`).

* `OpenTelemetry.Logs.LogRecord.ForEachScope<TState>(System.Action<OpenTelemetry.Logs.LogRecordScope, TState> callback, TState state)` became `OpenTelemetry.Logs.LogRecordStruct.ForEachScope<TState>(OpenTelemetry.Logs.LogRecordScopeCallback<TState> callback, TState state)`. Not breaking for users doing lambda ([example](https://github.com/open-telemetry/opentelemetry-dotnet/blob/d1c34ab235edc3816704c3a06fa8d3573efd9d9a/docs/logs/extending-the-sdk/MyExporter.cs#L49)). Breaking if using a strongly typed field ([example](https://github.com/open-telemetry/opentelemetry-dotnet/blob/d1c34ab235edc3816704c3a06fa8d3573efd9d9a/test/Benchmarks/Logs/LogScopeBenchmarks.cs#L31)). This had to be done due to the `readonly ref struct LogRecordScope` change above.

* Since `LogRecordScope` was being changed I removed `LogRecordScope.Enumerator` `Dispose` & `Reset` methods. They didn't do anything useful and removing `Dispose` allows the compiler to generate faster `foreach` code. Breaking only if `LogExporter` author is rolling their own `GetEnumerator`/`MoveNext`/`Current`/`Dispose` calls (not using `foreach`).

* Since `LogRecordScope` was being changed I made the ctor on `LogRecordScope` internal. Probably an oversight that it was exposed in the first place. Breaking if `LogExporter` author is newing `LogRecordScope.Enumerator` directly. No reason I can think of for why anyone would be doing that.

## TODOs

* [ ] Changes in public API reviewed
* [ ] `CHANGELOG` updates
* [ ] Unit tests