@cijothomas 2022-04-01
> @cijothomas @CodeBlanch any updates on this one? I saw that #2932 was closed.

The approach in *this* PR is much better than 2932, so we'll prefer this approach. This is somewhat big re-factoring, so will come after 1.2.